Extract audio control logic into hook
